1st

+++ Avalonia.NETCoreMVVMApp1/Avalonia.NETCoreMVVMApp1/Models/Dot.cs
@@ -2,7 +2,7 @@ namespace Avalonia.NETCoreMVVMApp1.Models
 {
     public class Dot
     {
-        public int X=0;
-        public int Y=0;
+        public double X =0;
+        public double Y=0;
     }
 }
